Foundation for GameMap, MapBoundary Physics
Added GameMap placeholder class for holding data about a map in the game
(just a shell with a MapBoundary at the moment). PhysicsSprite now
interacts with the MapBoundary rather than hardcoded values, but the
rules for the interaction are still hardcoded by the PhysicsSprite.
Added static reference to the model in the Game class; in the future it
could be a static list of all models currently instantiated.