MapBoundary Collisions Implementation
Physics for PhysicsSprite that collides with the map boundary
implemented, with the physics of the interaction controlled by the map
boundary. Objects colliding with the boundary bounce softly,
experiencing a force proportional to the distance outside the map
boundary and in the direction towards the inside of the boundary.

Moved the starting location for the player's ship so it will now start
inside the map boundary.