GameMap Framework
Several changes to serve as the future framework for more complex
GameMaps:

-GameMap now contains a list of only PhysicsSprites to speed up
collision detection loops. Multithreaded collision detection with
PhysicsSprites and MapBorder compatible but not optimized.
-A full list of PhysicsSprites on the map can no longer be directly
referenced by the PhysicsSprites. For collision detection, the full list
in the GameMap is passed to the PhysicsSprite as a parameter.
-Future framework for being able to create a blank map, a map from data
from a file, or a blank map.
-Mostly implementedcustom functions for adding sprites to lists, where
adding a sprite will also add the sprite to other relevant lists. (Some
public functions to get direct access to the list exist, but should NOT
be used to add/remove/change the list)