SpriteList and File Comments
Architecture changes with regards to the Sprite lists. SpriteList now
contains the master (static) list of all Sprites in the game. View uses
only this list to draw to the screen. Other classes such as GameMap will
add to this list in SpriteMap, and must keep track of any sprites added
to the list so it can remove them later.

General comments at the tops of files added, cleaned up, and updated.