GameGUI Support
Changes deal primarily with creating a general GUI class compatible with
the current work and phasing out the FullScreenMenu.

-Added better future support for a GUI with the GameGUI class. Old
functionality in the FullScreenMainMenu class has been absorbed into the
new class. One GameGUI object will exist across all game states, and
changing game states changes a change in appearance for the GUI
-Some changes to how Model and ModelVars interact may happen in the
future to clean up some data relationships
-Added a flag needsRedraw for sprite class to differentiate between when
a sprite needs to completely recreate its image or when the most
recently drawn image can be used
-Button pressed image currently broken. A proper fix may require
reworking the how the MenuText or general strings are drawn to the
screen