Early GUIWindow Implementation
GUIWindow partially implemented. GUIWindow shares similar functionality
with a GameGUI object, and it may be possible to use inheritance to
implement the class in the future.

-Added GUIWindow class. The GameGUI instantiates and removes GUIWindow
objects.
-Added OPEN_MENU and CLOSE_WINDOW as actions for ActionButtons
-GameGUI now has a single GUIWindow object. If it is null, there is no
window displayed on the screen. Instantiating the window makes a window
appear on the screen, and methods in the GUIWindow class will be used to
set what window is desired. (Currently only a very early pause game menu
is implemented)
-Minor change to text centering in MenuText to improve appearance, but
it may be adjusted again in the future
-Game now starts in a paused state with the pause menu pulled up. Game
can also be paused without opening the menu. If the menu is closed, the
game is automatically resumed.
-Added a working image for the GUIWindow background