Early Ingame Upgrades Window
Added functionality allowing the player to upgrade the ship through the
upgrade window. Still a work in progress; some GUI elements are not
functioning yet. Upgrades made through the GUI are reflected in the game

Major Changes:
-Added UPGRADE button ingame. Clicking the button allows one to access
the upcoming upgrade window.
-Added shield upgrade button to the upgrade window, where the ship can
be upgraded to a maximum of 3 (starting at 0).
-Added relevant ButtonController actions to allow for upgrading the
shields.
-Added variables to track the current upgrade level of the ship and the
relevant functions to access them.

Minor Changes:
-Added transparency to the MenuText shadow color, improving the
readability of the primary text
-Added a wide version of the gui window for the upgrades screen

Bugs:
-GUIWindow only updates when a mouse cursor hovers over the location of
a button